LG-12039: minor refactor and test.
