+++ spec/services/doc_auth/lexis_nexis/responses/true_id_response_spec.rb
@@ -431,6 +431,7 @@ RSpec.describe DocAuth::LexisNexis::Responses::TrueIdResponse do
     end
 
     it 'produces reasonable output for a malformed TrueID response' do
+      allow(NewRelic::Agent).to receive(:notice_error)
       output = described_class.new(failure_response_malformed, config).to_h
 
       expect(output[:success]).to eq(false)
